Basic scene, object, and component management.
Just getting started on the basic framework for the game. So far, I have
the ability to switch between scenes and customize the lifecycle of
different types of scenes. You can also add and remove GameObjects in a
scene, as well as add components to GameObjects, which manipulate their
behavior. I've decided to loosely follow Unity's API design, as it seems
intuitive and "just works." That's how I want my framework to behave.