Deprecated SafeList, more PhysicsBody functionality
SafeList was a great concept, but I think it was causing too much
overhead, as its enumerator is more complicated than that of a normal
list. If necessary, this may be added back in the future, but for now, a
copy of the list is iterated over, so the original list can still be
modified and the enumeration will not fail. The only difference is that
new objects will wait to be updated until the next update cycle, which
shouldn't be an issue.  Also exposed more PhysicsBody functionality - it
almost matches the original Body.