Added parent-child system.
I have implemented a parent-child system, so GameObjects can be parents
of other GameObjects.  This means that when the parent is deleted, all
of it's children are destroyed as well.  Right now, properties don't
carry over from parents to children, but this may (or many not) be
implemented in the future.