Added joints, property binding, texture rendering.
Added joints functionality; objects can now pivot relative to each
other. Also added property binding. The GameObject class used to have a
PhysicsBody instance that would be assigned to when a PhysicsBody
component was added. The position properties for the GameObject would
then refer to the position of the PhysicsBody. While this system did
work, it means PhyiscsBody gets extra rights, which kind of goes against
the system. So I added property binding, so any component can bind a
bindable GameObject property to one of its own. Also started on basic
texture rendering. Not much yet, but more to come soon.