Started on TMX loading.
Started the process of importing Tiled Map Editor files. I decided to do
this by creating a custom map importer and processor to optimize the
loading process. I'm essentially creating a code structure based on the
XML structure of .tmx files, serializing it, deserializing it, and using
it to generate the scene. So far so good.