Complete .tmx loading.
MonoEngine can now completely load a .tmx map, including objects in the
object layer.  There are a few holes, but those will be filled when the
time comes that a need for them to be filled arises.  Also added a
system where you can specify the type of object in the TMX map and that
type will be generated at runtime.