The commit might crash GitHub.
Well a lot changed. I couldn't move myself to make a commit between the
last one and this one because I was never fully satisfied with the state
of my code. I know, it's a lame excuse. But anyways. I changed the
entire way physics works with the game engine. Instead of manually
wrapping each aspect of the physics engine as I go, the engine now has
wrappers for the three main physics elements: Bodies, Joints, and
Fixtures. These three base wrappers now work for any physics component
derived from them. It also gives the developer more freedom with how
they do physics (which I need for this game). And then I started on the
ACTUAL GAME. It's about time. So far I have a test scene with the basic
map enclosure and a ramp in the middle. I have a car that has wheels and
drives on the ground. Oh, and on the walls too. And I have boosting and
flying. The car physics is rather complicated, and I'm trying to match
Rocket League's car physics as much as possible. I won't go into too
much detail here about how it works, but see Car.cs for the details (it
has some documentation so you at least know what the heck you're looking
at).